refactor(payment) update payment dtos

+++ backend/src/main/java/org/youcode/trackme/dtos/payment/PaymentRequestDTO.java
@@ -4,6 +4,6 @@ import org.youcode.trackme.entities.enums.PaymentStatus;
 
 public record PaymentRequestDTO(
 //        PaymentStatus status,
-        Double amount
+        Integer amount
 //        String transactionId
 ) {}
